correctly initializing calendar with selected value
