+++ applications/launchpad/gui-react/src/components/DatePicker/types.ts
@@ -1,0 +1,8 @@
+import { CSSProperties } from 'react'
+
+export type DatePickerProps = {
+  value?: Date
+  open: boolean
+  onChange: (d: Date) => void
+  style?: CSSProperties
+}
